ynyt/requirements.txt
fastparquet==2023.10.1
holidays==0.17.2
numpy==1.23.4
pandas==1.5.1
pandas==1.5.2
python_dateutil==2.8.2
s3fs==0.4.2
scikit_learn==0.24.1
scikit_learn==1.0.2
scipy==1.9.3
